+++ htdocs/photosynthesis/ajax-bill-notes.php
@@ -57,17 +57,20 @@ else
 {
 
     /*
-     * Clear the Memcached cache of comments on this bill, since Photosyntheis comments are
+     * Clear the Memcached cache of comments on this bill, since Photosynthesis comments are
      * among them.
      */
-    $sql = 'SELECT bill_id AS id
-			FROM dashboard_bills
-			WHERE id=' . $_POST['id'];
-    $result = mysql_query($sql);
-    $bill = mysql_fetch_array($result);
-    $mc = new Memcached();
-    $mc->addServer("127.0.0.1", 11211);
-    $mc->delete('comments-' . $bill['id']);
+    if (MEMCACHED_SERVER != '')
+    {
+        $sql = 'SELECT bill_id AS id
+                FROM dashboard_bills
+                WHERE id=' . $_POST['id'];
+        $result = mysql_query($sql);
+        $bill = mysql_fetch_array($result);
+        $mc = new Memcached();
+        $mc->addServer("127.0.0.1", 11211);
+        $mc->delete('comments-' . $bill['id']);
+    }
 
     echo $notes;
 }
